split Optimisation and BaseOptimsier classes, loosen standalone cost unit test
