+++ pybop/__init__.py
@@ -96,7 +96,7 @@ from ._experiment import Experiment
 #
 # Optimiser class
 #
-from .optimisers._optimisation import BaseOptimiser, Optimisation
+from .optimisers._optimisation import BaseOptimiser
 from .optimisers.base_optimiser import BasePintsOptimiser
 from .optimisers.scipy_optimisers import (
     BaseSciPyOptimiser,
@@ -113,6 +113,7 @@ from .optimisers.pints_optimisers import (
     SNES,
     XNES,
 )
+from .optimisers.optimisation import Optimisation
 
 #
 # Parameter classes
